requirements.txt
requests
proto-plus>=1.22.3
pydantic>=2.6.4
numpy
openpyxl
pymochow>=1.1.2
PyYAML
bce-python-sdk
pycryptodome
opentelemetry-exporter-otlp>=1.23.0
opentelemetry-instrumentation>=0.44b0
opentelemetry-sdk>=1.23.0
opentelemetry-api>=1.23.0
pydub>=0.10.0
ffmpeg
urllib3<2.0.0